+++ projects/Apps/common/README.md
@@ -1,0 +1,9 @@
+# Commo Code package
+
+This package is used as a store for common snippets of code used across mutliple package. This often refers to shared types that are used end to end
+
+## Test
+
+In the Apps folder run
+
+`yarn workspace @guardian/editions-common run test`
