+++ projects/Apps/common/src/collection/card-layouts.ts
