+++ projects/Apps/common/src/collection/layout-model.ts
