+++ projects/Apps/common/src/helpers.ts
