+++ projects/Mallard/src/components/front/image-resource.tsx
@@ -9,7 +9,7 @@ import {
 } from 'react-native'
 import { useAspectRatio } from 'src/hooks/use-aspect-ratio'
 import { useImagePath, useScaledImage } from 'src/hooks/use-image-paths'
-import { Image as IImage, ImageUse } from '../../../../common/src'
+import { Image as IImage, ImageUse } from '../../../../Apps/common/src'
 
 /**
  * This component abstracts away the endpoint for images
