+++ projects/Mallard/src/components/front/items/trail-image-view.tsx
@@ -6,7 +6,7 @@ import { useMediaQuery } from 'src/hooks/use-screen'
 import { Breakpoints } from 'src/theme/breakpoints'
 import { ImageResource } from '../image-resource'
 import { SportScore } from 'src/components/sportscore/sportscore'
-import { ArticleType } from '../../../../../common/src'
+import { ArticleType } from '../../../../../Apps/common/src'
 
 const trailImageViewStyles = StyleSheet.create({
     frame: {
