+++ projects/Mallard/src/helpers/files.ts
@@ -3,7 +3,7 @@ import RNFetchBlob from 'rn-fetch-blob'
 import { Issue } from 'src/common'
 import { getIssueSummary } from 'src/hooks/use-issue-summary'
 import { FSPaths } from 'src/paths'
-import { ImageSize, IssueSummary } from '../../../common/src'
+import { ImageSize, IssueSummary } from '../../../Apps/common/src'
 import { lastSevenDays, todayAsKey } from './issues'
 import { imageForScreenSize } from './screen'
 import { getSetting } from './settings'
