+++ projects/Mallard/src/hooks/use-image-paths.ts
@@ -3,7 +3,7 @@ import ImageResizer from 'react-native-image-resizer'
 import RNFetchBlob from 'rn-fetch-blob'
 import { imageForScreenSize } from 'src/helpers/screen'
 import { APIPaths, FSPaths } from 'src/paths'
-import { Image, ImageSize, Issue, ImageUse } from '../../../common/src'
+import { Image, ImageSize, Issue, ImageUse } from '../../../Apps/common/src'
 import { useIssueSummary } from './use-issue-summary'
 import { Platform } from 'react-native'
 import { useApiUrl } from './use-settings'
