+++ projects/Mallard/src/screens/issue-screen.tsx
@@ -47,7 +47,7 @@ import { metrics } from 'src/theme/spacing'
 import { useIssueScreenSize, WithIssueScreenSize } from './issue/use-size'
 import { useQuery } from 'src/hooks/apollo'
 import gql from 'graphql-tag'
-import { IssueWithFronts, Front as TFront } from '../../../common/src'
+import { IssueWithFronts, Front as TFront } from '../../../Apps/common/src'
 import {
     flattenCollectionsToCards,
     flattenFlatCardsToFront,
