+++ projects/backend/common.ts
@@ -1,3 +1,3 @@
-export * from '../common/src/index'
-export * from '../common/src/helpers'
+export * from '../Apps/common/src/index'
+export * from '../Apps/common/src/helpers'
 export * from '../backend/utils/try'
