--- projects/common/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "@guardian/editions-common",
-    "@notes": "metro bundler needs babel here to not crap out :(.",
-    "dependencies": {
-        "@babel/runtime": "^7.5.5",
-        "@types/ramda": "^0.26.18",
-        "ramda": "^0.26.1"
-    },
-    "devDependencies": {
-      "typescript": "^3.5.3",
-      "@types/jest": "^24.0.17",
-      "jest": "^24.8.0",
-      "ts-jest": "^24.0.2"
-    },
-    "scripts": {
-        "build": "echo 'Common does not need building.'",
-        "test": "jest"
-    }
-}
